部署脚本优化

+++ deploy/run.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 PRO_NAME=$1
-pid=$(cat "$PRO_NAME".pid)
-echo "结束进程 $pid"
-kill "$pid"
-
+if [ -f "$PRO_NAME.pid" ]; then
+    pid=$(cat "$PRO_NAME".pid)
+    if [ ! -z "$pid" ]; then
+        echo "结束进程 $pid"
+        kill "$pid"
+    fi
+fi
 nohup ./"$PRO_NAME" -mode=prod >"$PRO_NAME".log 2>&1 &
 echo -e "\n"
 echo "nohup success"
